+++ src/Component.hs
@@ -2,7 +2,9 @@
 -- | Root module.
 --   Provides all things required to use and compose components.
 
-module Component (module Component.Class, module Component.List) where
+module Component (module Component.Class, module Component.List, module Component.Find) where
 
 import Component.Class
 import Component.List
+import Component.Find
+